jax: read param_decomp_config directly; delete torch_config.py
Fold the wrapper-yaml -> ExperimentConfig conversion (formerly the standalone
torch_config.py "conversion layer") into config.py, which now reads the canonical
param_decomp_config schema (LMExperimentConfig) DIRECTLY. No separate parallel
representation module. convert_torch_lm_config -> build_experiment_config,
load_torch_wrapper -> load_wrapper; the site-resolution code is moved byte-for-byte
so the resolved set/order is preserved (#725/#728/#763). Rewire run.py / export.py /
tests; rename test_torch_config.py -> test_config.py. The `torch_config:` wrapper-yaml
key (canonical in param_decomp_config.jax_wrapper.WRAPPER_KEYS) is unchanged.

Closes #806
