+++ param_decomp_jax/jax_single_pool/export.py
@@ -40,7 +40,7 @@ from safetensors.numpy import save_file
 
 from jax_single_pool.checkpoint import make_checkpoint_manager, restore_step
 from jax_single_pool.ci_fn import CIFn
-from jax_single_pool.config import TargetConfig
+from jax_single_pool.config import TargetConfig, load_run_dir_config
 from jax_single_pool.llama8b import (
     KIND_ORDER,
     DecompVU,
@@ -54,7 +54,6 @@ from jax_single_pool.llama8b import (
 from jax_single_pool.lm import SiteSpec
 from jax_single_pool.run_state import build_optimizers, init_train_state
 from jax_single_pool.sharding import dp_mesh
-from jax_single_pool.torch_config import load_run_dir_config
 
 CI_FN_PREFIX = "ci_fn._global_ci_fn"
